flushing input buffer for serial

+++ minions/Arduino_DHT11_12xLEDs/Arduino_DHT11_12xLEDs.ino
@@ -14,7 +14,7 @@ DHT dht(DHTPIN, DHTTYPE);
 CRGB leds[NUMOFLEDS];
 String led_mode = "off";
 
-float temperature, humidity;
+double temperature, humidity;
 int water_low = 0;
 
 void setup() {
